+++ angular-workspace/projects/ni/nimble-angular/CHANGELOG.md
@@ -1,8 +1,32 @@
 # Change Log - @ni/nimble-angular
 
-This log was last generated on Wed, 23 Nov 2022 16:47:02 GMT and should not be manually modified.
+This log was last generated on Wed, 07 Dec 2022 13:10:43 GMT and should not be manually modified.
 
 <!-- Start content -->
+
+## 14.0.9
+
+Wed, 07 Dec 2022 13:10:43 GMT
+
+### Patches
+
+- Bump @ni/nimble-components to v15.5.3
+
+## 14.0.8
+
+Mon, 05 Dec 2022 19:44:44 GMT
+
+### Patches
+
+- Bump @ni/nimble-components to v15.5.2
+
+## 14.0.7
+
+Tue, 29 Nov 2022 17:43:36 GMT
+
+### Patches
+
+- Bump @ni/nimble-components to v15.5.1
 
 ## 14.0.6